Fix main function in compressor

+++ src/f32_noa_comp_gpu.cu
@@ -378,7 +378,7 @@ int main(int argc, char* argv [])
     dtimer.start();
 
     d_reset<<<1, 1>>>();
-    cudaMemset(d_fullcarry, 0, chunks * sizeof(byte));
+    cudaMemset(d_fullcarry, 0, chunks * sizeof(int));
 
     thrust::device_ptr<float> dev_ptr = thrust::device_pointer_cast((float*)d_input);
     thrust::pair<thrust::device_ptr<float>, thrust::device_ptr<float>> min_max = thrust::minmax_element(thrust::device, dev_ptr, dev_ptr + (insize / sizeof(float)));
